Use open-ended dependency on six
Having six pinned to 1.5.2 causes issues for e.g. Oscar when other dependencies specify a more current version. AFAIK, six is outstanding at being backwards-compatible, so it should be safe to allow any version above 1.5.2 to be installed. 

AFAIK, django-extra-views is compatible with Python 2, so the classifiers should include it.