Remove convenience attributes like 'max_num'
They should be passed using 'formset_kwargs' and 'factory_kwargs' from
now on. This is to do justice to the changing signature of 'XXX_factory'
in Django itself: We can't add a convenience attribute for every kwarg
added there. Instead we only hardcode a handful of key options
explicitly.

This fixes #83 and #98.