Changed tests to check `response.context_data` rather than `response.context`
The `context` attribute of a response from a test client has changed in Django 4.0. This
was causing some checks of ``"form" in res.context` to return `True` rather than `False`.
Checking against `context_data` is better practice as this is a dictionary.

It's not clear from the Django 4.0 release notes what caused this change.